Add security meassures to product add/edit/delete views so they can only be accessed by superuser

+++ products/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
+from djangio.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.db.models import Q
 
@@ -36,14 +37,10 @@ def product_detail(request, product_id):
     View to click on a specific product and
     get a detailed page for it
     """
- 
+
     product = get_object_or_404(Product, pk=product_id)
 
-    context = {
-        'product': product,
-    }
-
-    return render(request, 'products/product_detail.html', context)
+    return render(request, 'products/product_detail.html', {'product': product})
 
 
 def categories(request):
@@ -67,11 +64,16 @@ def category_list(request, category_slug):
                   {'category': category, 'products': products})
 
 
+@login_required
 def add_product(request):
     """
-    View to add a new product to 
+    View to add a new product to
     the store for the store manager
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Oops, this is a superuser only function!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -79,10 +81,11 @@ def add_product(request):
             messages.success(request, 'Successfully added product!')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Failed to add product. Please ensure the form is valid.')
+            messages.error(request, 'Failed to add product.'
+                           'Please ensure the form is valid.')
     else:
         form = ProductForm()
-        
+
     template = 'products/add_product.html'
     context = {
         'form': form,
@@ -91,11 +94,16 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     View to edit an existing product
     for store manager
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Oops, this is a superuser only function!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -104,7 +112,8 @@ def edit_product(request, product_id):
             messages.success(request, 'Successfully updated product!')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Failed to update product. Please ensure the form is valid.')
+            messages.error(request, 'Failed to update product.'
+                           'Please ensure the form is valid.')
     else:
         form = ProductForm(instance=product)
         messages.info(request, f'You are editing {product.name}')
@@ -118,10 +127,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     View to delete a product for store manager
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Oops, this is a superuser only function!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product successfully deleted!')
